% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDPdata}
\alias{GDPdata}
\title{1960-2003 GDP dataset}
\format{
A data frame with 748 rows and 14 columns. For detailed explanations of the exact measures, see \url{https://www.dropbox.com/s/n1bvqb54xrw8vyj/GPSynth.pdf?dl=0}:
\describe{
  \item{Index}
  \item{Country}
  \item{Year}
  \item{GDP}
  \item{Infrate}
  \item{Trade}
  \item{Schooling}
  \item{Invest60}
  \item{Invest70}
  \item{Invest80}
  \item{Industry}
  \item{Invest} 
  \item{School}
  \item{Ind}
  }
}
\source{
\url{https://www.rottentomatoes.com}
}
\usage{
GDPdata
}
\description{
An example dataset for using \code{\ink{GPP}} to estimate the counterfactual GDP of West Germany assuming no reunification.
}
\seealso{
\code{\link{GPP.R}} \code{\link{plotGPPfit.R}} \code{\link{writeMod.R}} \code{\link{runMod.R}} \code{\link{autoconverge.R}}
}
\keyword{datasets}
