% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMod.R
\name{writeMod}
\alias{writeMod}
\alias{writeMod,ANY-method}
\title{Writes Stan code for GPP model}
\usage{
writeMod(noise, ncov, printMod)
}
\arguments{
\item{noise}{The desired amount of artificial noise to add to the model.}

\item{ncov}{The number of covariates to include in the model.}
}
\value{
A string of Stan code that can be run with \code{\link{runMod.R}}
}
\description{
Returns string of Stan code that can be run to estimate the GPP.
}
\examples{

writeMod(noise = 0.25, ncov = 2)

}
\seealso{
\code{\link{plotGPPfit.R}} \code{\link{runMod.R}} \code{\link{GPP.R}} \code{\link{autoconverge.R}}
}
\author{
Devin P. Brown \email{devinpbrown96@gmail.com} and David Carlson \email{carlson.david@wustl.edu}
}
